requirements.txt
Flask
gunicorn
requests
python-dotenv
numpy
pandas
scikit-learn>=1.0.2
statsmodels>=0.13.2
matplotlib>=3.5.1
tensorflow>=2.9.0
torch>=1.11.0
transformers>=4.18.0
langchain>=0.0.267
openai>=0.27.0
seaborn>=0.11.2
plotly>=5.6.0
dash>=2.3.0
keras>=2.9.0
xgboost>=1.5.1
optuna>=2.10.0
reinforce>=0.1.0
streamlit>=1.8.1
faiss-cpu>=1.7.2
chromadb>=0.3.21
